+++ src/Curiosity.Migrations.PostgreSQL/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Improved cancellation support in `PostgresMigrationConnection`.
+- Improved parsing `PostgresException` in `MigrationActionHelper`.
 
 ## [4.0.2] - 2023-04-10